extract ErrorResponse
