+++ src/main/java/ru/mail/park/exceptions/ErrorResponse.java
@@ -1,0 +1,27 @@
+package ru.mail.park.exceptions;
+
+public class ErrorResponse {
+    private String msg;
+    private String error;
+
+    public ErrorResponse(String error, String msg) {
+        this.msg = msg;
+        this.error = error;
+    }
+
+    public String getMsg() {
+        return msg;
+    }
+
+    public void setMsg(String msg) {
+        this.msg = msg;
+    }
+
+    public String getError() {
+        return error;
+    }
+
+    public void setError(String error) {
+        this.error = error;
+    }
+}
