+++ src/main/java/ru/mail/park/exceptions/ExceptionHandlerAdvice.java
@@ -19,29 +19,4 @@ public class ExceptionHandlerAdvice {
         return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new ErrorResponse(HttpStatus.BAD_REQUEST.toString(), "Server error"));
     }
 
-    private static final class ErrorResponse {
-        private String msg;
-        private String error;
-
-        private ErrorResponse(String error, String msg) {
-            this.msg = msg;
-            this.error = error;
-        }
-
-        public String getMsg() {
-            return msg;
-        }
-
-        public void setMsg(String msg) {
-            this.msg = msg;
-        }
-
-        public String getError() {
-            return error;
-        }
-
-        public void setError(String error) {
-            this.error = error;
-        }
-    }
 }
